cud product

+++ src/main/java/com/example/demo/service/impl/ProductServiceImpl.java
@@ -32,18 +32,28 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProductById(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new RuntimeException("product with "+id+" not found");
+        }
         productRepository.deleteById(id);
     }
 
     @Override
-    public void updateProduct(Long id, Product product) {
-        Optional<Product> optionalProduct = productRepository.findById(id);
-        Product product1 = optionalProduct.get();
+    public Product updateProduct(Product product) {
+        Optional<Product> optionalProduct = productRepository.findById(product.getId());
+        Product productToUpdate = optionalProduct.orElseThrow(
+                ()-> new RuntimeException("Product with id" + product.getId() + " does not exist")
+        );
+
+        updateProduct(productToUpdate, product);
+        return productRepository.save(productToUpdate);
+
+
     }
 
     @Override
     public Product addProduct(Product product) {
-        return null;
+        return productRepository.save(product);
     }
 
     @Override
@@ -51,5 +61,33 @@ public class ProductServiceImpl implements ProductService {
         List<Product> products = new ArrayList<>();
         products = productRepository.findByCategory_Id(categoryId);
         return products;
+    }
+
+    private void updateProduct(Product productToUpdate, Product product) {
+        if (product.getId() != null) {
+            productToUpdate.setId(product.getId());
+        }
+        if (product.getName() != null) {
+            productToUpdate.setName(product.getName());
+        }
+        if (product.getDescription() != null) {
+            productToUpdate.setDescription(product.getDescription());
+        }
+        if (product.getPrice() != null) {
+            productToUpdate.setPrice(product.getPrice());
+        }
+        if (product.getDiscount() != null) {
+            productToUpdate.setDiscount(product.getDiscount());
+        }
+        if (product.getImageUrl() != null) {
+            productToUpdate.setImageUrl(product.getImageUrl());
+        }
+        if (product.getCategory() != null) {
+            productToUpdate.setCategory(product.getCategory());
+        }
+
+        if (product.getSize()!=null) {
+            productToUpdate.setSize(product.getSize());
+        }
     }
 }
